+++ CHANGELOG.md
@@ -1,6 +1,37 @@
 # CHANGELOG
 
 
+## v0.58.1 (2025-08-26)
+
+### Bug Fixes
+
+- **cli**: Include plugin template test package
+  ([#967](https://github.com/canvas-medical/canvas-plugins/pull/967),
+  [`3e40473`](https://github.com/canvas-medical/canvas-plugins/commit/3e40473cf0f7251f3269c9aa904a87ce93f37c65))
+
+
+## v0.58.0 (2025-08-26)
+
+### Features
+
+- **cli**: Add local database support to the run-plugin cli command
+  ([#859](https://github.com/canvas-medical/canvas-plugins/pull/859),
+  [`8c6417a`](https://github.com/canvas-medical/canvas-plugins/commit/8c6417acdaf6b11cdb6d2e96ce169adf6810d10d))
+
+- **tests**: Add test utilities package
+  ([#842](https://github.com/canvas-medical/canvas-plugins/pull/842),
+  [`ba7f406`](https://github.com/canvas-medical/canvas-plugins/commit/ba7f406a62844432103d740ea74e02737e2b2bcf))
+
+
+## v0.57.1 (2025-08-25)
+
+### Bug Fixes
+
+- Set size maximums for gRPC messages
+  ([#914](https://github.com/canvas-medical/canvas-plugins/pull/914),
+  [`82cc22d`](https://github.com/canvas-medical/canvas-plugins/commit/82cc22d834da4420986cfed87f570471c8ff4973))
+
+
 ## v0.57.0 (2025-08-24)
 
 ### Chores